fix: rename HoursCard to HoursTable
This aligns better with the PagesComponent underlying component

Requested by Aaron
